feat(api): add Flask-Migrate for database migrations
This commit adds Flask-Migrate to the project for managing database
migrations. The `Migrate` object is initialized in `app.py` and
`env.py` is added to the `migrations` directory for running migrations.

A new migration script is also added to the `migrations/versions`
directory. This script alters the `name` column of the `company` table
from a `VARCHAR(64)` to a `VARCHAR(255)`.

Note: The migration script has an empty commit message.